requirements pipreqs . --force.txt
click==8.2.1
fastapi==0.115.14
joblib==1.4.2
joblib==1.5.1
numpy==2.3.1
pandas==2.3.0
passlib==1.7.4
pydantic==2.11.7
python-dotenv==1.1.1
python_jose==3.5.0
Requests==2.32.4
scikit_learn==1.7.0
